refactor(wc3): encapsulate Raven Form and locomotion in abilities
