+++ games/warcraft-3/game/skills/s_move.c
@@ -10,12 +10,9 @@
  * the unit has arrived it switches to the stand (idle) state; otherwise it
  * rotates toward the goal and advances by one frame's worth of movement.
  *
- * Collision separation between walking units is handled in g_phys.c
- * (G_SolveCollisions) after all units have moved.
+ * Steering, collision-aware steps, route goals, and support heights are owned here.
  */
 #include "s_skills.h"
-
-void move_walk(LPEDICT ent);
 
 /* With move-time collision (block-and-slide), "blocked" now means the unit
  * could not take a step this frame because it was boxed in — common and
@@ -32,6 +29,763 @@ typedef struct {
     VECTOR2 point;
     FLOAT radius;
 } moveSlot_t;
+
+#define MOVE_SLIDE_STEP BZ_ROUTE_SLIDE_STEP
+#define MOVE_SLIDE_RINGS BZ_ROUTE_SLIDE_RINGS
+#define MOVE_SLIDE_RINGS_YIELD 2                               /* +/- 30 deg: faster unit holds its line */
+#define MOVE_WORKER_QUEUE_TICKS 4                              /* same-stream blocker: queue before passing */
+#define MOVE_WORKER_ESCAPE_TICKS 8                             /* widen bounded escape corridor after this */
+#define MOVE_WORKER_CORRIDOR_RESET (30.0f * (FLOAT)M_PI / 180.0f)
+#define MOVE_WORKER_MAX_DEVIATION 5.0f                         /* collision radii */
+#define MOVE_WORKER_ESCAPE_DEVIATION 6.0f                      /* collision radii */
+#define MAX_MOVE_COLLIDERS     256
+
+typedef enum {
+    MOVE_AVOID_GENERIC,
+    MOVE_AVOID_RESOURCE_WORKER,
+    MOVE_AVOID_STATIC_ONLY,
+} moveAvoidPolicy_t;
+
+typedef enum {
+    MOVE_COLLIDE_UNITS,
+    MOVE_IGNORE_UNITS,
+} moveCollisionPolicy_t;
+
+static LPEDICT trymove_self = NULL;
+static LPEDICT trymove_blocker = NULL;  /* unit that rejected the last candidate (NULL = clear or terrain) */
+static LPEDICT trymove_colliders[MAX_MOVE_COLLIDERS];
+
+/* Wrap an angle delta into [-PI, PI]. */
+static FLOAT angle_wrap(FLOAT a) {
+    while (a > (FLOAT)M_PI)  a -= 2.0f * (FLOAT)M_PI;
+    while (a < -(FLOAT)M_PI) a += 2.0f * (FLOAT)M_PI;
+    return a;
+}
+
+/* unit_changeangle is defined lower down — it needs the move-validity test and
+ * the give-way helpers, which are declared below. */
+
+/* A unit is actively executing a ground move order (right-click move). */
+BOOL unit_is_walking(LPCEDICT ent) {
+    return ent->currentmove && ent->currentmove->ability == &a_move;
+}
+
+/* Location orders own a legal ground endpoint; interaction orders route to an
+ * entity centre and let their behavior-specific range decide arrival. */
+static BOOL unit_routes_to_location(LPCEDICT ent) {
+    if (!ent->currentmove)
+        return false;
+    if (ent->currentmove->ability == &a_move || ent->currentmove->ability == &a_patrol)
+        return true;
+    return ent->currentmove->ability == &a_attack && ent->goalentity == ent->movement.attackmove_waypoint;
+}
+
+/* Unit's effective current move speed.  Group moves travel at the slowest
+ * member's speed so the selection stays a cohesive formation instead of
+ * stringing out (WC3); the cap is gated on the move state so it never leaks
+ * into a later attack/harvest order that reuses this.  Using the *capped*
+ * speed means members of one group compare equal (no give-way within a group). */
+static FLOAT unit_current_speed(LPCEDICT self) {
+    FLOAT speed = self->unitinfo.MoveSpeed > 0
+        ? self->unitinfo.MoveSpeed
+        : self->data.UnitBalance->speed;
+    if (self->movement.group_speed > 0 && self->movement.group_speed < speed && unit_is_walking(self)) {
+        speed = self->movement.group_speed;
+    }
+    return speed;
+}
+
+FLOAT unit_movedistance(LPEDICT self) {
+    return 10 * unit_current_speed(self) / FRAMETIME;
+}
+
+/* --- Collision-aware movement (block-and-slide) ---------------------------
+ *
+ * A unit only commits a step into a position that is free of walkable terrain
+ * and of other units' collision circles.  When the steered heading is blocked
+ * it tries progressively larger left/right deflections ("sliding"), so units
+ * flow around obstacles instead of plowing through them.  Idle units are hard,
+ * immovable obstacles: walking into one never displaces it (the WC3 invariant
+ * that the old post-move push solver violated). */
+
+static BOOL unit_is_flying(LPCEDICT ent) {
+    return (ent->aiflags & AI_FLYING) != 0;
+}
+
+/* BoxEdicts predicate: solid units/buildings sharing this mover's collision
+ * layer.  Excludes self, hollow entities, zero-collision entities (waypoints,
+ * effects, missiles), and the opposite air/ground layer (flyers and ground
+ * units pass through each other). */
+static BOOL filter_blockers(LPCEDICT ent) {
+    if (ent == trymove_self || IS_HOLLOW(ent) || ent->collision <= 0.0f)
+        return false;
+    /* An alive walkable destructable is a ground surface, not a circle-shaped
+     * obstacle. Its authored path texture remains responsible for deck edges. */
+    if (G_IsDestructable(ent) && !ent->destructable.dead &&
+        ent->destructable.placement_solid && ent->pathtex &&
+        ent->data.DestructableData && ent->data.DestructableData->walkable) return false;
+    /* Trees have collisionSize 0 (they block only via their baked footprint) so
+     * they are already excluded above; buildings keep a real collision circle
+     * and ARE counted here — relying on the terrain footprint alone let units
+     * walk through buildings (coarse 32u cells, runtime-spawned statics not yet
+     * baked).  Flyers and ground units are on separate layers. */
+    return unit_is_flying(ent) == unit_is_flying(trymove_self);
+}
+
+/* Distance from point p to the segment [a,b]. */
+static FLOAT point_segment_distance(LPCVECTOR2 a, LPCVECTOR2 b, LPCVECTOR2 p) {
+    VECTOR2 const ab = Vector2_sub(b, a);
+    VECTOR2 const ap = Vector2_sub(p, a);
+    FLOAT const ab2 = ab.x * ab.x + ab.y * ab.y;
+    FLOAT t = ab2 > 0.0001f ? (ap.x * ab.x + ap.y * ab.y) / ab2 : 0.0f;
+    if (t < 0.0f) t = 0.0f;
+    else if (t > 1.0f) t = 1.0f;
+    VECTOR2 const closest = { a->x + t * ab.x, a->y + t * ab.y };
+    return Vector2_distance(&closest, p);
+}
+
+/* Is the position 'cand' free for 'self' (static world + other units)?  On a
+ * unit rejection, records the blocking unit in trymove_blocker (NULL otherwise)
+ * so the slide can apply speed-priority give-way. */
+static BOOL move_is_valid_policy(LPEDICT self, LPCVECTOR2 cand,
+                                 moveCollisionPolicy_t collision_policy) {
+    trymove_blocker = NULL;
+    /* Pathing-disabled units (SetUnitPathing(false), scripted moves) ignore
+     * all collision, matching the old unconditional translate. */
+    if (self->no_pathing)
+        return true;
+
+    /* Static world: terrain + baked building footprints (pathmap.original). */
+    if (!CM_PointIsPathableForRadius(cand, self->collision))
+        return false;
+    /* WC3's pathing grid rejects a swept step that cuts a diagonal corner. Keep
+     * the escape case for units spawned inside stale/changed pathing, where the
+     * endpoint remains the authoritative legal position. */
+    if (CM_PointIsPathableForRadius(&self->s.origin2, self->collision) &&
+        !CM_LineIsWalkableForRadius(&self->s.origin2, cand, self->collision))
+        return false;
+
+    if (collision_policy == MOVE_IGNORE_UNITS)
+        return true;
+
+    /* Dynamic units: precise circle test.  The "don't deepen penetration" rule
+     * ignores a neighbour the unit already overlaps unless the candidate moves
+     * closer to it, so units that start overlapped (spawn / blink / a building
+     * dropped on them) can still slide apart instead of dead-locking. */
+    /* Broad-phase box must cover the whole swept segment (origin -> cand), not
+     * just the endpoint: a fast unit's step spans many units, and a box centred
+     * on cand would miss a blocker sitting near the START of the path — letting
+     * the unit jump clean over it.  BoxEdicts tests each entity's bounds (which
+     * already extend by its own collision radius), so inflating by self's radius
+     * is enough to catch any blocker within rr of the corridor. */
+    FLOAT const reach = self->collision + 1.0f;
+    FLOAT const ox = self->s.origin2.x, oy = self->s.origin2.y;
+    BOX2 const box = {
+        { (ox < cand->x ? ox : cand->x) - reach, (oy < cand->y ? oy : cand->y) - reach },
+        { (ox > cand->x ? ox : cand->x) + reach, (oy > cand->y ? oy : cand->y) + reach },
+    };
+    trymove_self = self;
+    DWORD const num = gi.BoxEdicts(&box, trymove_colliders, MAX_MOVE_COLLIDERS, filter_blockers);
+    FOR_LOOP(i, num) {
+        LPEDICT const b = trymove_colliders[i];
+        FLOAT const rr = self->collision + b->collision;
+        /* Swept test: the unit's whole PATH this tick (origin -> cand) must
+         * clear b, not just the endpoint — otherwise a fast unit (step ~one
+         * cell) jumps clean over a smaller unit between ticks.  Mirrors WC3's
+         * swept-circle collision. */
+        FLOAT const seg_d = point_segment_distance(&self->s.origin2, cand, &b->s.origin2);
+        if (seg_d >= rr)
+            continue;  /* the swept path clears b */
+        FLOAT const cur_d = Vector2_distance(&self->s.origin2, &b->s.origin2);
+        if (cur_d < rr && seg_d >= cur_d - 0.5f)
+            continue;  /* already overlapping b: allow only a step whose path does
+                        * not go deeper into b — lets it separate, never slide
+                        * tangentially or jump THROUGH it. */
+        trymove_blocker = b;
+        return false;
+    }
+    return true;
+}
+
+static BOOL move_is_valid(LPEDICT self, LPCVECTOR2 cand) {
+    return move_is_valid_policy(self, cand, MOVE_COLLIDE_UNITS);
+}
+
+/* Shared steering uses the same static-only policy as resource interaction movement. */
+static BOOL move_static_is_valid(LPEDICT self, LPCVECTOR2 cand) { return move_is_valid_policy(self, cand, MOVE_IGNORE_UNITS); }
+
+/* Public: would 'pos' be a free standing spot for 'self' (terrain + units)?
+ * Used by the move arrival to avoid snapping a unit onto an occupied goal. */
+BOOL M_MoveIsValid(LPEDICT self, LPCVECTOR2 pos) {
+    return move_is_valid(self, pos);
+}
+
+static void unit_commit_step(LPEDICT self, LPCVECTOR2 cand) {
+    if (self->s.flags & EF_FOW_BLOCKER) G_FowMarkBlockersDirty();
+    self->s.origin2 = *cand;
+    gi.LinkEntity(self);
+}
+
+/* Advance the unit one tick.  Avoidance is decided ONCE per tick in
+ * unit_changeangle (which picks a free heading via unit_desired_heading and
+ * turns the facing toward it); this function only commits the step.  WC3 moves a
+ * unit ALONG ITS FACING, so we try the facing first; if the facing momentarily
+ * lags into an obstacle while it is still turning toward the chosen heading, we
+ * fall back to that already-validated heading so the unit keeps progressing
+ * around the obstacle instead of stalling.
+ *
+ * We deliberately do NOT run a second deflection search here.  The previous
+ * version searched +/- slide rings off the (turn-rate-lagged) facing, which
+ * disagreed with the heading unit_changeangle had already chosen and re-decided
+ * a different direction every tick — that disagreement is what made units
+ * visibly rotate/wobble and crab sideways past each other and trees. */
+static void unit_moveindirection_policy(LPEDICT self,
+                                        moveCollisionPolicy_t collision_policy) {
+    if (self->aiflags & AI_IMMOBILE)
+        return;
+
+    /* unit_changeangle* clears both routing fields before resolving this
+     * tick's heading.  A resumable cache miss deliberately leaves both clear;
+     * in that state there is no valid movement decision yet.  Never commit a
+     * step using the unit's previous facing/heading while the requested route
+     * is still being built.  This is the common safety net for Move, Harvest,
+     * Patrol, Attack, Build, Repair, and resource-return walkers. */
+    if (!self->movement.flow_direct && !self->movement.path.valid && self->movement.flow_generation == 0)
+        return;
+
+    FLOAT const dist = unit_movedistance(self);
+    VECTOR2 const by_facing = Vector2_mad(&self->s.origin2, dist,
+                                          &MAKE(VECTOR2, cosf(self->s.angle), sinf(self->s.angle)));
+    if (move_is_valid_policy(self, &by_facing, collision_policy)) {
+        unit_commit_step(self, &by_facing);
+        return;
+    }
+    VECTOR2 const by_heading = Vector2_mad(&self->s.origin2, dist,
+                                           &MAKE(VECTOR2, cosf(self->movement.heading), sinf(self->movement.heading)));
+    if (move_is_valid_policy(self, &by_heading, collision_policy)) {
+        unit_commit_step(self, &by_heading);
+    }
+}
+
+void unit_moveindirection(LPEDICT self) {
+    unit_moveindirection_policy(self, MOVE_COLLIDE_UNITS);
+}
+
+void unit_moveindirection_ignore_units(LPEDICT self) {
+    unit_moveindirection_policy(self, MOVE_IGNORE_UNITS);
+}
+
+/* Interaction routing may finish at a collision-safe staging point rather
+ * than at the blocked building centre.  When that endpoint is within this
+ * tick's movement budget, land exactly on it instead of stepping past it and
+ * selecting it again from the opposite side next think.  This is the same
+ * arrival snap used by ordinary Move, but deliberately ignores live units for
+ * Warsmash-style Mine/drop-off legs while retaining all static pathing. */
+BOOL unit_snap_to_point_ignore_units(LPEDICT self, LPCVECTOR2 point) {
+    if (!self || !point || (self->aiflags & AI_IMMOBILE))
+        return false;
+    if (Vector2_distance(&self->s.origin2, point) > unit_movedistance(self) + 0.001f)
+        return false;
+    if (!move_is_valid_policy(self, point, MOVE_IGNORE_UNITS))
+        return false;
+    unit_commit_step(self, point);
+    return true;
+}
+
+/* Turn the facing vector toward a target heading by at most the unit's turn
+ * rate ('umvr', radians/tick; WC3 default 0.5).  Pure 2-D vector math (cross =
+ * signed sin of the angle to turn, dot = cos); atan2 only writes the canonical
+ * s.angle the renderer/network consume. */
+static void unit_turn_toward(LPEDICT self, FLOAT target) {
+    VECTOR2 const facing = { cosf(self->s.angle), sinf(self->s.angle) };
+    VECTOR2 const goal   = { cosf(target), sinf(target) };
+    FLOAT const cross = facing.x * goal.y - facing.y * goal.x;
+    FLOAT const dot   = facing.x * goal.x + facing.y * goal.y;
+    FLOAT turn = self->data.UnitData->turnRate;
+    if (turn <= 0.0f) turn = 0.5f;
+
+    if (dot >= cosf(turn)) {
+        self->s.angle = target;  /* within one tick's turn: snap */
+    } else {
+        FLOAT const st = cross >= 0.0f ? sinf(turn) : -sinf(turn);
+        FLOAT const ct = cosf(turn);
+        VECTOR2 const nf = { facing.x * ct - facing.y * st,
+                             facing.x * st + facing.y * ct };
+        self->s.angle = atan2f(nf.y, nf.x);
+    }
+}
+
+/* Resource workers need a different local crowd rule from ordinary combat
+ * movement.  A same-direction worker is a queue, not an obstacle to weave
+ * around; crossing traffic may pass immediately.  This is the minimal policy
+ * that stayed close to the direct Human02 resource corridor in the 30-worker
+ * simulation while still breaking counterflow deadlocks. */
+static BOOL unit_worker_same_stream(LPCEDICT blocker, FLOAT goal_angle) {
+    VECTOR2 dir, goal;
+    FLOAT len;
+
+    if (!blocker || !blocker->currentmove || !blocker->goalentity ||
+        (blocker->aiflags & AI_IMMOBILE))
+        return false;
+    dir = Vector2_sub(&blocker->goalentity->s.origin2, &blocker->s.origin2);
+    len = Vector2_len(&dir);
+    if (len <= 0.001f)
+        return false;
+    goal = MAKE(VECTOR2, cosf(goal_angle), sinf(goal_angle));
+    return Vector2_dot(&goal, &dir) / len > 0.25f;
+}
+
+static FLOAT unit_worker_lateral_deviation(LPCEDICT self, LPCVECTOR2 point) {
+    VECTOR2 const delta = Vector2_sub(point, &self->movement.worker_avoid_origin);
+    VECTOR2 const direct = { cosf(self->movement.worker_avoid_heading),
+                             sinf(self->movement.worker_avoid_heading) };
+    return fabsf(direct.x * delta.y - direct.y * delta.x);
+}
+
+static FLOAT unit_worker_desired_heading(LPEDICT self, FLOAT goal_angle, FLOAT dist) {
+    VECTOR2 const straight = Vector2_mad(&self->s.origin2, dist,
+                                         &MAKE(VECTOR2, cosf(goal_angle), sinf(goal_angle)));
+    LPEDICT blocker;
+    FLOAT max_deviation;
+
+    if (move_is_valid(self, &straight)) {
+        self->movement.worker_avoid_blocked_frames = 0;
+        self->movement.worker_avoid_active = false;
+        return goal_angle;
+    }
+
+    blocker = trymove_blocker;
+    if (!self->movement.worker_avoid_active ||
+        fabsf(angle_wrap(goal_angle - self->movement.worker_avoid_heading)) >
+            MOVE_WORKER_CORRIDOR_RESET) {
+        self->movement.worker_avoid_origin = self->s.origin2;
+        self->movement.worker_avoid_heading = goal_angle;
+        self->movement.worker_avoid_blocked_frames = 0;
+        self->movement.worker_avoid_active = true;
+    }
+    self->movement.worker_avoid_blocked_frames++;
+
+    /* Do not turn a short pause in a resource stream into overtaking.  Four
+     * blocked decisions let the queue advance naturally; a genuinely pinned
+     * queue then gets the same bounded escape used for crossing traffic. */
+    if (unit_worker_same_stream(blocker, goal_angle) &&
+        self->movement.worker_avoid_blocked_frames <= MOVE_WORKER_QUEUE_TICKS)
+        return goal_angle;
+
+    max_deviation = self->collision *
+        (self->movement.worker_avoid_blocked_frames <= MOVE_WORKER_ESCAPE_TICKS
+            ? MOVE_WORKER_MAX_DEVIATION : MOVE_WORKER_ESCAPE_DEVIATION);
+
+    /* Deterministic right-hand passing avoids the +/- re-decision that made
+     * packed Peasants dance.  Retry the exact direct heading next think; no
+     * passing lane is cached. */
+    for (int sign = -1; sign <= 1; sign += 2) {
+        for (int ring = 1; ring <= MOVE_SLIDE_RINGS; ring++) {
+            FLOAT const angle = angle_wrap(goal_angle + sign * ring * MOVE_SLIDE_STEP);
+            VECTOR2 const cand = Vector2_mad(&self->s.origin2, dist,
+                                             &MAKE(VECTOR2, cosf(angle), sinf(angle)));
+            if (unit_worker_lateral_deviation(self, &cand) > max_deviation)
+                continue;
+            if (move_is_valid(self, &cand)) {
+                self->movement.worker_avoid_blocked_frames = 0;
+                return angle;
+            }
+        }
+    }
+    return goal_angle;
+}
+
+/* Pick the heading the unit actually wants to move along this tick.  Generic
+ * units retain speed-priority block-and-slide; resource workers use the
+ * queue/pass-right policy above. */
+static FLOAT unit_desired_heading(LPEDICT self, FLOAT goal_angle, FLOAT dist,
+                                  moveAvoidPolicy_t policy) {
+    moveCollisionPolicy_t const collision_policy =
+        policy == MOVE_AVOID_STATIC_ONLY ? MOVE_IGNORE_UNITS : MOVE_COLLIDE_UNITS;
+    VECTOR2 const straight = Vector2_mad(&self->s.origin2, dist,
+                                         &MAKE(VECTOR2, cosf(goal_angle), sinf(goal_angle)));
+    if (policy == MOVE_AVOID_RESOURCE_WORKER)
+        return unit_worker_desired_heading(self, goal_angle, dist);
+    if (move_is_valid_policy(self, &straight, collision_policy))
+        return goal_angle;
+
+    int max_rings = MOVE_SLIDE_RINGS;
+    LPEDICT const b = trymove_blocker;
+    if (b && unit_is_walking(self) && unit_is_walking(b) &&
+        unit_current_speed(self) > unit_current_speed(b)) {
+        max_rings = MOVE_SLIDE_RINGS_YIELD;
+    }
+    ROUTESLIDE slide = { .ent = self, .angle = goal_angle, .dist = dist, .rings = max_rings,
+        .valid = collision_policy == MOVE_IGNORE_UNITS ? move_static_is_valid : move_is_valid };
+    return CM_SlideRoute(&slide);
+}
+
+static void unit_apply_heading(LPEDICT self, LPCVECTOR2 dir, moveAvoidPolicy_t policy) {
+    FLOAT const dirlen = Vector2_len(dir);
+    if (dirlen <= 0.001f)
+        return;  /* no meaningful heading this tick: hold current facing */
+
+    /* Local avoidance resolves into ONE heading; the facing turns toward it and
+     * the move step (unit_moveindirection) follows it, keeping facing and motion
+     * aligned (no second, disagreeing search). */
+    FLOAT const goal_angle = atan2f(dir->y, dir->x);
+    FLOAT const desired = unit_desired_heading(self, goal_angle,
+                                                unit_movedistance(self), policy);
+    self->movement.heading = desired;
+    unit_turn_toward(self, desired);
+}
+
+static void unit_changeangle_towards_point_policy(LPEDICT self, LPCVECTOR2 point,
+                                                   moveAvoidPolicy_t policy) {
+    VECTOR2 dir;
+
+    if (!self || !point || (self->aiflags & AI_IMMOBILE))
+        return;
+    self->movement.heading = self->s.angle;
+    self->movement.flow_generation = 0;
+    self->movement.flow_goal_reached = false;
+    self->movement.flow_unreachable = false;
+    self->movement.flow_direct = true;
+    dir = Vector2_sub(point, &self->s.origin2);
+    unit_apply_heading(self, &dir, policy);
+}
+
+/* Keep the bounded point-route turn until it is reached; retail likewise owns
+ * route progress on each mover instead of rebuilding from its current point. */
+static BOOL unit_accel_direction_to_point(LPEDICT self, LPCVECTOR2 target,
+                                          FLOAT radius, LPVECTOR2 dir) {
+    if (!self || !target || !dir) return false;
+    pathAccelParams_t params = { &self->s.origin2, target, radius };
+    return CM_AccelerateRoute(&self->movement.path, &params, dir);
+}
+
+static BOOL unit_accel_direction(LPEDICT self, FLOAT radius, LPVECTOR2 dir) {
+    return unit_accel_direction_to_point(self, &self->goalentity->s.origin2,
+                                         radius, dir);
+}
+
+void unit_changeangle_towards_point(LPEDICT self, LPCVECTOR2 point) {
+    unit_changeangle_towards_point_policy(self, point, MOVE_AVOID_GENERIC);
+}
+
+void unit_changeangle_towards_point_worker(LPEDICT self, LPCVECTOR2 point) {
+    unit_changeangle_towards_point_policy(self, point, MOVE_AVOID_RESOURCE_WORKER);
+}
+
+BOOL unit_changeangle_towards_point_ignore_units(LPEDICT self, LPCVECTOR2 point) {
+    VECTOR2 dir;
+
+    if (!self || !point || (self->aiflags & AI_IMMOBILE))
+        return false;
+
+    self->movement.heading = self->s.angle;
+    self->movement.flow_generation = 0;
+    self->movement.flow_goal_reached = false;
+    self->movement.flow_unreachable = false;
+    self->movement.flow_direct = false;
+
+    /* Resource-return behaviors keep the building as their authoritative goal,
+     * but navigation may target a worker-relative footprint edge.  Prefer that
+     * exact point when it is directly reachable; otherwise use the same
+     * collision-sized mover-owned A* accelerator used while shared fields are
+     * pending.  Live units remain ignored by the steering/move policy. */
+    if (CM_LineIsWalkableForRadius(&self->s.origin2, point, self->collision)) {
+        self->movement.path.valid = false;
+        self->movement.flow_direct = true;
+        dir = Vector2_sub(point, &self->s.origin2);
+    } else if (!unit_accel_direction_to_point(self, point, self->collision, &dir)) {
+        return false;
+    }
+
+    unit_apply_heading(self, &dir, MOVE_AVOID_STATIC_ONLY);
+    return true;
+}
+
+static void unit_changeangle_policy(LPEDICT self, moveAvoidPolicy_t policy) {
+    if (self->aiflags & AI_IMMOBILE)
+        return;
+    VECTOR2 to_goal = Vector2_sub(&self->goalentity->s.origin2, &self->s.origin2);
+    VECTOR2 dir;
+    FLOAT const radius = unit_routes_to_location(self) ? self->collision : 0.0f;
+
+    self->movement.heading = self->s.angle;  /* default if no heading is resolved this tick */
+    self->movement.flow_generation = 0;
+    self->movement.flow_goal_reached = false;
+    self->movement.flow_unreachable = false;
+    self->movement.flow_direct = false;
+
+    /* Generic interaction movement keeps the original point-route contract.
+     * Attack, mine entry, resource return, repair, and other ranged behaviors
+     * decide when their interaction boundary has been reached.  Do not stop
+     * those orders at a collision-expanded flow goal outside that boundary.
+     * Move orders own radius-valid reserved destinations, so their route must
+     * use the same footprint as move-time collision; point routing previously
+     * sent units into narrow gaps and touching obstacle corners. */
+    if (CM_LineIsWalkableForRadius(&self->s.origin2, &self->goalentity->s.origin2, radius)) {
+        self->movement.path.valid = false;
+        self->movement.flow_direct = true;
+        dir = to_goal;
+    } else {
+        DWORD heatmap = M_RefreshHeatmap(self->goalentity, radius);
+        self->movement.flow_generation = heatmap;
+        if (!heatmap) {
+            if (!unit_accel_direction(self, radius, &dir))
+                return; /* long incremental route is still building; keep the order */
+            /* path_valid resolves the heading while the shared field builds;
+             * this is not a direct line to the requested destination. */
+            unit_apply_heading(self, &dir, policy);
+            return;
+        }
+        self->movement.path.valid = false;
+        if (CM_FlowReachedGoal(heatmap, self->s.origin.x, self->s.origin.y)) {
+            /* Location orders stop at their collision-safe route endpoint in
+             * the owning behavior.  Interaction orders use a point field whose
+             * raw target may be blocked (mine/building/unit centre); once the
+             * adjusted route end is reached they must steer toward the real
+             * target so the behavior's footprint/range check can complete. */
+            self->movement.flow_goal_reached = true;
+            dir = to_goal;
+        } else {
+            dir = get_flow_direction(heatmap, self->s.origin.x, self->s.origin.y);
+            if (Vector2_len(&dir) <= 0.001f) {
+                self->movement.flow_unreachable = !CM_FlowCanReach(heatmap, self->s.origin.x, self->s.origin.y);
+                /* Location targets are private waypoints.  When the clicked static
+                 * component is unreachable, replace the waypoint with the
+                 * closest legal point in this mover's component; aiming at the
+                 * raw click made local avoidance walk forever along walls. */
+                if (radius > 0.0f && self->movement.flow_unreachable) {
+                    VECTOR2 closest;
+                        if (CM_ClosestReachablePointForRadius(
+                            &self->s.origin2, &self->goalentity->s.origin2, radius, &closest)) {
+                        self->goalentity->s.origin2 = closest;
+                        self->goalentity->secondarygoal = NULL;
+                        self->goalentity->heatmap2 = 0;
+                        self->goalentity->heatmap2_radius = 0;
+                        move_reset_progress(self);
+                    }
+                    return;
+                }
+                return;
+            }
+        }
+    }
+
+    unit_apply_heading(self, &dir, policy);
+}
+
+void unit_changeangle(LPEDICT self) {
+    unit_changeangle_policy(self, MOVE_AVOID_GENERIC);
+}
+
+void unit_changeangle_worker(LPEDICT self) {
+    unit_changeangle_policy(self, MOVE_AVOID_RESOURCE_WORKER);
+}
+
+/* Behaviors that route around authored blocked geometry may request a
+ * collision-sized field. Lumber uses its route-end state to retarget an
+ * unreachable tree. Build and Repair instead route toward behavior-owned legal
+ * approach points, so reaching the adjusted flow goal never changes their
+ * gameplay target. Generic point movement continues through unit_changeangle(). */
+static void unit_changeangle_for_radius_policy(LPEDICT self, FLOAT radius,
+                                               moveAvoidPolicy_t policy,
+                                               BOOL continue_to_target) {
+    if (self->aiflags & AI_IMMOBILE)
+        return;
+    VECTOR2 to_goal = Vector2_sub(&self->goalentity->s.origin2, &self->s.origin2);
+    VECTOR2 dir;
+
+    self->movement.heading = self->s.angle;
+    self->movement.flow_generation = 0;
+    self->movement.flow_goal_reached = false;
+    self->movement.flow_unreachable = false;
+    self->movement.flow_direct = false;
+
+    if (CM_LineIsWalkableForRadius(&self->s.origin2,
+                                   &self->goalentity->s.origin2,
+                                   radius)) {
+        self->movement.path.valid = false;
+        self->movement.flow_direct = true;
+        dir = to_goal;
+    } else {
+        DWORD heatmap = M_RefreshHeatmap(self->goalentity, radius);
+        self->movement.flow_generation = heatmap;
+        if (!heatmap) {
+            if (!unit_accel_direction(self, radius, &dir))
+                return; /* long incremental route is still building */
+            unit_apply_heading(self, &dir, policy);
+            return;
+        }
+        self->movement.path.valid = false;
+
+        if (CM_FlowReachedGoal(heatmap, self->s.origin.x, self->s.origin.y)) {
+            self->movement.flow_goal_reached = true;
+            if (!continue_to_target)
+                return;
+            /* Ranged interactions target a blocked unit/building centre.  A
+             * collision-sized route deliberately ends at the nearest legal
+             * cell around that footprint; from there keep steering at the real
+             * target and let the behavior's precise range check complete the
+             * interaction before a step would enter static pathing. */
+            dir = to_goal;
+        } else {
+            dir = get_flow_direction(heatmap, self->s.origin.x, self->s.origin.y);
+        }
+        if (!self->movement.flow_goal_reached && Vector2_len(&dir) <= 0.001f) {
+            self->movement.flow_unreachable =
+                !CM_FlowCanReach(heatmap, self->s.origin.x, self->s.origin.y);
+            return;
+        }
+    }
+
+    unit_apply_heading(self, &dir, policy);
+}
+
+void unit_changeangle_for_radius(LPEDICT self, FLOAT radius) {
+    unit_changeangle_for_radius_policy(self, radius, MOVE_AVOID_GENERIC, false);
+}
+
+void unit_changeangle_for_radius_worker(LPEDICT self, FLOAT radius) {
+    unit_changeangle_for_radius_policy(self, radius, MOVE_AVOID_RESOURCE_WORKER, false);
+}
+
+void unit_changeangle_interaction_ignore_units(LPEDICT self) {
+    /* Warsmash's mover owns a collision-sized path even when the ranged
+     * behavior disables unit collision.  Use the worker radius for static
+     * routing, but keep mobile-unit collision disabled at steering/move time. */
+    unit_changeangle_for_radius_policy(self, self ? self->collision : 0.0f,
+                                       MOVE_AVOID_STATIC_ONLY, true);
+}
+
+/* Reserve a body-queue-style ring in g_edicts so ordinary F_EDICT relocation owns every waypoint pointer. */
+void G_InitWaypoints(void) {
+    DWORD base;
+    if (level.waypoints.count) return;
+    base = level.waypoints.base = globals.num_edicts;
+    FOR_LOOP(i, MAX_WAYPOINTS) {
+        LPEDICT waypoint = G_Spawn();
+        if (waypoint != g_edicts + base + i) gi.error("G_InitWaypoints: waypoint ring is not contiguous\n");
+        waypoint->svflags |= SVF_NOCLIENT;
+    }
+    level.waypoints.count = MAX_WAYPOINTS;
+}
+
+/* Recycle one real edict from the fixed ring, matching Quake II's TRAIL/body queue ownership model. */
+LPEDICT Waypoint_add(LPCVECTOR2 spot) {
+    LPEDICT waypoint;
+    G_InitWaypoints();
+    waypoint = g_edicts + level.waypoints.base + level.waypoints.cursor;
+    level.waypoints.cursor = (level.waypoints.cursor + 1) % MAX_WAYPOINTS;
+    waypoint->s.origin.x = spot->x;
+    waypoint->s.origin.y = spot->y;
+    waypoint->heatmap2 = 0;
+    waypoint->heatmap2_radius = 0;
+    waypoint->secondarygoal = NULL;
+    waypoint->collision = 0;
+    M_CheckGround(waypoint);
+    return waypoint;
+}
+
+static int move_harvest_path_debug_level(void) {
+    LPCSTR value;
+    value = gi.CvarString("wc3_harvest_path_debug", "0");
+    return value ? atoi(value) : 0;
+}
+
+DWORD M_RefreshHeatmap(LPEDICT self, FLOAT radius) {
+    LPEDICT route = self && self->secondarygoal ? self->secondarygoal : self;
+    BOOL radius_matches;
+    BOOL cached = false;
+    DWORD generation;
+
+    if (!route)
+        return 0;
+
+    radius_matches = fabsf(route->heatmap2_radius - radius) < 0.01f;
+    if (radius_matches && route->heatmap2)
+        cached = CM_ActivateCachedFlow(route->heatmap2);
+
+    /* Fixed waypoints never move, so a still-cached field remains valid until
+     * static pathing invalidates the routing cache. */
+    if (cached && !(route->svflags & SVF_MONSTER))
+        return route->heatmap2;
+
+    if (cached && (route->svflags & SVF_MONSTER)) {
+        BOOL const moved = Vector2_distance(&route->s.origin2, &route->heatmap2_origin) >= 64.0f;
+        BOOL const stale = (DWORD)(level.time - route->heatmap2_time) >= 400;
+        if (!moved || !stale)
+            return route->heatmap2;
+    }
+
+    /* Cache misses are resumable in common/routing.c.  Return the old field for
+     * a moving target while its replacement is being built; fixed goals with
+     * no field simply wait until a later tick instead of steering straight into
+     * the obstacle that caused routing to be needed. */
+    generation = CM_RequestHeatmapForRadius(route, radius);
+    if (!generation)
+        return cached ? route->heatmap2 : 0;
+
+    route->heatmap2 = generation;
+    route->heatmap2_origin = route->s.origin2;
+    route->heatmap2_time = level.time;
+    route->heatmap2_radius = radius;
+
+    if (move_harvest_path_debug_level() >= 2 && route->targtype == TARG_TREE) {
+        fprintf(stderr,
+                "WC3_HARVEST_PATH heatmap target=%d reason=ready generation=%u radius=%.1f\n",
+                route->s.number, route->heatmap2, radius);
+    }
+    return route->heatmap2;
+}
+
+static LPCSTR M_UnitMoveTypeName(LPCEDICT self) {
+    return self && self->data.UnitData ? self->data.UnitData->moveTypeName : NULL;
+}
+
+static BOOL M_UnitUsesWaterSurface(LPCEDICT self, LPCSTR movetp) {
+    if (!movetp) return false;
+    if (!strcmp(movetp, "fly") || !strcmp(movetp, "hover") || !strcmp(movetp, "float"))
+        return true;
+    if (!strcmp(movetp, "amph")) {
+        return CM_TerrainPointIsSwimmable(&self->s.origin2) &&
+               !CM_TerrainPointIsWalkable(&self->s.origin2);
+    }
+    return false;
+}
+
+/* Resolve the visual/support surface, then apply the unit's mutable fly height.
+ * FOOT/HORSE stay terrain-based; FLY/HOVER/FLOAT and swimming AMPH units use
+ * max(terrain, water).  Walkable destructables can raise every movement type
+ * except FLOAT, matching Warsmash's "boats can't go on bridges" rule. */
+void M_CheckGround(LPEDICT self) {
+    LPCSTR const movetp = M_UnitMoveTypeName(self);
+    BOOL const floating = movetp && !strcmp(movetp, "float");
+    FLOAT height = CM_GetHeightAtPoint(self->s.origin.x, self->s.origin.y);
+    FLOAT const cell = CM_PathCellWorldSize();
+
+    if (M_UnitUsesWaterSurface(self, movetp))
+        height = MAX(height, CM_GetWaterHeightAtPoint(self->s.origin.x, self->s.origin.y));
+
+    if (!floating) {
+        for (LPEDICT surface = level.ground_surfaces; surface; surface = surface->ground_next) {
+            pathTex_t const *pathtex = surface->pathtex;
+            if (!surface->inuse || surface->destructable.dead ||
+                !surface->destructable.placement_solid || !pathtex) continue;
+            if (fabsf(self->s.origin.x - surface->s.origin.x) > pathtex->width * cell * 0.5f ||
+                fabsf(self->s.origin.y - surface->s.origin.y) > pathtex->height * cell * 0.5f) continue;
+            height = MAX(height, surface->s.origin.z);
+        }
+    }
+    self->s.ground_offset = self->unitinfo.FlyHeight;
+    self->s.origin.z = height + self->s.ground_offset;
+}
+
+FLOAT M_DistanceToGoal(LPEDICT ent) {
+    if (ent->goalentity) {
+        return Vector2_distance(&ent->goalentity->s.origin2, &ent->s.origin2);
+    } else {
+        return 0;
+    }
+}
 
 static FLOAT move_slot_spacing(LPEDICT const *units, DWORD count) {
     FLOAT max_radius = 0;
