+++ games/warcraft-3/game/skills/s_raven.c
@@ -1,0 +1,184 @@
+#include "s_skills.h"
+
+static void raven_forward_end(LPEDICT unit);
+static void raven_reverse_end(LPEDICT unit);
+
+static umove_t raven_morph = { "morph", NULL, raven_forward_end, &a_raven_form };
+static umove_t raven_morph_alt = { "morph alternate", NULL, raven_reverse_end, &a_raven_form };
+static LPCSTR const raven_orders[] = { "ravenform", "unravenform", NULL };
+
+typedef struct ravenform_s {
+    AbilityData_t const *ability;
+    DWORD base_type;
+    DWORD raven_type;
+} RAVENFORM;
+typedef RAVENFORM *LPRAVENFORM;
+typedef RAVENFORM const *LPCRAVENFORM;
+
+static DWORD const raven_codes[] = {
+    MAKEFOURCC('A','m','r','f'), /* Medivh Crow Form */
+    MAKEFOURCC('A','r','a','v'), /* Druid Storm Crow Form */
+};
+
+/* Play the authored morph sequence after a form rebind; the reverse transform
+ * uses the source form's Alternate sequence before returning to ordinary stand. */
+static void raven_play_morph(LPEDICT unit, BOOL raven_form) {
+    LPCANIMATION morph;
+
+    if (raven_form) G_AddUnitAnimationProperties(unit, "alternate,alternateex", false);
+    /* Keep the morph as the active move, rather than only replacing the
+     * animation pointer; the ordinary stand move can otherwise reassert its
+     * animation while the geoset alpha track is revealing the new form. */
+    unit_setmove(unit, raven_form ? &raven_morph : &raven_morph_alt);
+    morph = unit->animation;
+    if (raven_form) {
+        G_AddUnitAnimationProperties(unit, "alternateex", true);
+        /* Keep the untagged Morph selected after restoring the form tag;
+         * later cinematic orders must immediately use crow animations. */
+        unit->animation = morph;
+    }
+    if (!morph) {
+        fprintf(stderr, "WC3_RAVEN missing morph animation order=%s class=%.4s\n",
+                raven_form ? "ravenform" : "unravenform", (LPCSTR)&unit->class_id);
+        /* TODO: Custom models may omit morph clips; complete the form after reporting the missing sequence. */
+        (raven_form ? raven_forward_end : raven_reverse_end)(unit);
+    }
+    if (unit->animation) unit->s.frame = unit->animation->interval[0];
+}
+
+/* Takeoff is independent of the active order once the destination form has been installed. */
+static void raven_begin_rise(LPEDICT unit) {
+    if (unit->raven.rise_state != RAVEN_RISE_AFTER_MORPH) return;
+    if (unit->raven.rise_duration > 0.0f) {
+        unit->raven.rise_start = (FLOAT)G_Time();
+        unit->raven.rise_state = RAVEN_RISE_ACTIVE;
+    } else {
+        unit->unitinfo.FlyHeight = unit->raven.fly_height;
+        unit->raven.rise_state = RAVEN_RISE_NONE;
+        M_CheckGround(unit);
+        gi.LinkEntity(unit);
+    }
+}
+
+static void raven_forward_end(LPEDICT unit) {
+    /* nmdm requires alternateex; alternate would select Medivh's base stand sequence. */
+    G_AddUnitAnimationProperties(unit, "alternateex", true);
+    raven_begin_rise(unit);
+    unit_stand(unit);
+}
+
+/* Advance Raven Form's authored takeoff height independently from animation. */
+static void raven_update(LPEDICT unit) {
+    FLOAT fraction;
+    if (!unit) return;
+    /* Prologue01 replaces Morph with Move after 0.5s. Its endfunc then never runs;
+     * finish the pending takeoff transition without replacing the new order. */
+    if (unit->raven.rise_state == RAVEN_RISE_AFTER_MORPH && unit->currentmove != &raven_morph)
+        raven_begin_rise(unit);
+    if (unit->raven.rise_state != RAVEN_RISE_ACTIVE) return;
+    fraction = ((FLOAT)G_Time() - unit->raven.rise_start) / (unit->raven.rise_duration * 1000.0f);
+    if (fraction >= 1.0f) {
+        unit->unitinfo.FlyHeight = unit->raven.fly_height;
+        unit->raven.rise_state = RAVEN_RISE_NONE;
+    } else {
+        unit->unitinfo.FlyHeight = unit->raven.fly_height * MAX(0.0f, fraction);
+    }
+    M_CheckGround(unit);
+    gi.LinkEntity(unit);
+}
+
+static void raven_reverse_end(LPEDICT unit) {
+    G_AddUnitAnimationProperties(unit, "alternate,alternateex", false);
+    unit_stand(unit);
+}
+
+/* Preplaced campaign forms resolve through the same authored endpoints as learned abilities. */
+static BOOL raven_form_data(LPEDICT unit, LPRAVENFORM out) {
+    if (!unit || !out) return false;
+    FOR_LOOP(i, sizeof(raven_codes) / sizeof(raven_codes[0])) {
+        AbilityData_t const *ability = G_AbilityData(raven_codes[i]);
+        DWORD const base_type = ability->level[0].data[0].id;
+        DWORD const raven_type = ability->level[0].unitID;
+        RAVENFORM current;
+
+        if (!ability->id || !base_type || !raven_type) continue;
+        current = (RAVENFORM){
+            .ability = ability,
+            .base_type = base_type,
+            .raven_type = raven_type,
+        };
+        /* Endpoint identity is authoritative for preplaced campaign forms,
+         * which may not expose the transform ability through the runtime skill
+         * list before the map issues unravenform. */
+        if (unit->class_id == base_type || unit->class_id == raven_type) {
+            *out = current;
+            return true;
+        }
+    }
+    return false;
+}
+
+/* Rebind in place so script handles and selection survive both morph directions. */
+static BOOL raven_form_order(LPEDICT unit, BOOL raven_form) {
+    RAVENFORM form = {0};
+    DWORD target_type;
+
+    if (!raven_form_data(unit, &form)) {
+        fprintf(stderr, "Raven Form: no transform endpoints for %.4s\n", (LPCSTR)&unit->class_id);
+        return false;
+    }
+    target_type = raven_form ? form.raven_type : form.base_type;
+
+    if (unit->class_id == target_type) {
+        return true;
+    }
+    if (raven_form ? unit->class_id != form.base_type : unit->class_id != form.raven_type) {
+        return false;
+    }
+
+    G_ClearUnitOrderQueue(unit);
+    if (!G_TransformUnitType(unit, target_type)) {
+        fprintf(stderr, "Raven Form: cannot transform %.4s to %.4s\n", (LPCSTR)&unit->class_id, (LPCSTR)&target_type);
+        return false;
+    }
+
+    unit->raven.rise_state = RAVEN_RISE_NONE;
+    if (raven_form) {
+        unit->raven.fly_height = unit->unitinfo.FlyHeight;
+        unit->raven.rise_duration = form.ability->level[0].data[2].number;
+        unit->raven.rise_state = RAVEN_RISE_AFTER_MORPH;
+        unit->unitinfo.FlyHeight = 0.0f;
+        M_CheckGround(unit);
+        gi.LinkEntity(unit);
+    }
+
+    unit->goalentity = NULL;
+    unit->secondarygoal = NULL;
+    move_reset_progress(unit);
+    unit_stand(unit);
+    raven_play_morph(unit, raven_form);
+    return true;
+}
+
+/* Orders and the command card share the same ability-owned form transition. */
+static BOOL raven_order(LPEDICT unit, LPCSTR order) {
+    return raven_form_order(unit, !strcmp(order, raven_orders[0]));
+}
+
+static BOOL raven_is_on(LPEDICT unit) {
+    RAVENFORM form;
+    return raven_form_data(unit, &form) && unit->class_id == form.raven_type;
+}
+
+static void raven_command(LPEDICT ent) {
+    FOR_CONTROLLABLE_SELECTED_UNITS(ent->client, unit)
+        unit_issueimmediateorder(unit, raven_orders[raven_is_on(unit) ? 1 : 0]);
+}
+
+ability_t a_raven_form = {
+    .cmd = raven_command,
+    .is_toggle_on = raven_is_on,
+    .orders = raven_orders,
+    .order = raven_order,
+    .update = raven_update,
+};
