+++ games/warcraft-3/game/tests/t_game.c
@@ -2829,6 +2829,10 @@ SAVE_INT_FIELD_TEST(field_avatar_level_round_trip, avatar.level, 2)
 SAVE_INT_FIELD_TEST(field_avatar_damage_round_trip, avatar.damage, 31)
 SAVE_FLOAT_FIELD_TEST(field_avatar_armor_round_trip, avatar.armor, 7.0f)
 SAVE_FLOAT_FIELD_TEST(field_avatar_health_round_trip, avatar.health, 600.0f)
+SAVE_FLOAT_FIELD_TEST(field_raven_height_round_trip, raven.fly_height, 125.0f)
+SAVE_FLOAT_FIELD_TEST(field_raven_start_round_trip, raven.rise_start, 1000.0f)
+SAVE_FLOAT_FIELD_TEST(field_raven_duration_round_trip, raven.rise_duration, 2.0f)
+SAVE_INT_FIELD_TEST(field_raven_state_round_trip, raven.rise_state, RAVEN_RISE_ACTIVE)
 SAVE_INT_FIELD_TEST(field_polymorph_ability_round_trip, polymorph.ability, MAKEFOURCC('A', 'p', 'l', 'y'))
 SAVE_INT_FIELD_TEST(field_polymorph_buff_round_trip, polymorph.buff, MAKEFOURCC('B', 'p', 'l', 'y'))
 SAVE_INT_FIELD_TEST(field_polymorph_form_type_round_trip, polymorph.form_type, MAKEFOURCC('o', 'p', 'e', 'o'))
